Add files via upload
An extension of the data cleaning project done on the employee layoff dataset. I explored the cleaned dataset. I worked on grouping the sums of employees laid off by companies and industries. I also analyzed the average percentage of laid-off employees and grouped the values by the companies, and even by the countries.